+++ R/plotAcousticScene.R
@@ -23,6 +23,8 @@
 #'   if \code{FALSE} then only outlines will be plotted
 #' @param alpha transparency percentage for plotting, values less than 1
 #'   will allow multiple overlapping colors to be seen
+#' @param returnData if \code{TRUE} then no plot will be generated, instead the
+#'   dataframe that would normally be used to make the plot will be returned
 #' @param add logical flag if \code{FALSE} plots normally if \code{TRUE}
 #'   then the output can be (maybe) added to an existing ggplot object
 #'
@@ -52,7 +54,7 @@ plotAcousticScene <- function(x, freqMap, typeCol='species',
                               title=NULL, bin='1day', scale=c('log', 'linear'),
                               freqMin=NULL, freqMax=NULL,
                               fill=TRUE,
-                              alpha=1, add=FALSE) {
+                              alpha=1, returnData=FALSE, add=FALSE) {
     x <- checkSimple(x, needCols=c('UTC', typeCol))
     x$plotStart <- floor_date(x$UTC, unit=bin)
     thisPeriod <- unitToPeriod(bin)
@@ -89,7 +91,9 @@ plotAcousticScene <- function(x, freqMap, typeCol='species',
         d$group <- NULL
         d
     }))
-
+    if(isTRUE(returnData)) {
+        return(x)
+    }
     if(is.null(freqMin)) {
         freqMin <- min(x[['freqMin']])
     }